shared values for python tools
