+++ build.py
@@ -9,27 +9,7 @@ import json
 import os
 import sys
 
-PANDA_PATH = "./pandas"
-OUTPUT_PATH = "./export/redpanda.json"
-ZOO_PATH = "./zoos" 
-
-class DateConsistencyError(ValueError):
-    pass
-
-class DateFormatError(ValueError):
-    pass
-
-class GenderFormatError(ValueError):
-    pass
-
-class IdError(KeyError):
-    pass
-
-class LinkError(IndexError):
-    pass
-
-class NameFormatError(ValueError):
-    pass
+from shared import *
 
 class RedPandaGraph:
     """Class with the redpanda database and format/consistency checks.
